refactor: UserCard 리팩토링
- SearchedUserProps 타입 제거
- 구조 변경
